+++ changelog.md
@@ -14,6 +14,7 @@ All notable changes to this project will be documented in this file. The format 
 - New addObserver() and removeObserver() methods to Window class that allows users to listen for state changes made to the applications window.
 - New split() method to Font class which works like wrap() but provides a list of strings instead of a single formatted one.
 - New runGarbageCollection terminal command that can be used to reclaim heap memory at runtime.
+- New terminal commands listMonitors and listSpeakers for exposing available peripheral devices to developers.
 
 ### Changed
 - Fixed bug that wouldn't update the value of virtual gamepads when Input.setVirtualGamepadInput() was called.